fix(cli): address review comments
