+++ crates/iota/tests/cli_tests.rs
@@ -12,7 +12,6 @@ use std::{
     fmt::Write,
     fs::{self, read_dir},
     io::{self, Read, Seek, SeekFrom, Write as IoWrite},
-    net::SocketAddr,
     path::{Path, PathBuf},
     str, thread,
     time::Duration,
@@ -27,7 +26,7 @@ use iota::{
         SwitchResponse, TxProcessingArgs, estimate_gas_budget,
     },
     client_ptb::ptb::{PTB, PTBCommandResult},
-    iota_commands::{IotaCommand, IotaEnvConfig, parse_host_port},
+    iota_commands::{IotaCommand, IotaEnvConfig},
     key_identity::{KeyIdentity, get_identity_address},
 };
 use iota_config::IOTA_CLIENT_CONFIG;
@@ -5033,39 +5032,6 @@ async fn test_clever_errors() -> Result<(), anyhow::Error> {
 
     insta::assert_snapshot!(error_string);
     Ok(())
-}
-
-#[tokio::test]
-async fn test_parse_host_port() {
-    let input = "127.0.0.0";
-    let result = parse_host_port(input.to_string(), 9123).unwrap();
-    assert_eq!(result, "127.0.0.0:9123".parse::<SocketAddr>().unwrap());
-
-    let input = "127.0.0.5:9124";
-    let result = parse_host_port(input.to_string(), 9123).unwrap();
-    assert_eq!(result, "127.0.0.5:9124".parse::<SocketAddr>().unwrap());
-
-    let input = "9090";
-    let result = parse_host_port(input.to_string(), 9123).unwrap();
-    assert_eq!(result, "0.0.0.0:9090".parse::<SocketAddr>().unwrap());
-
-    let input = "";
-    let result = parse_host_port(input.to_string(), 9123).unwrap();
-    assert_eq!(result, "0.0.0.0:9123".parse::<SocketAddr>().unwrap());
-
-    let result = parse_host_port("localhost".to_string(), 9899).unwrap();
-    assert_eq!(result, "127.0.0.1:9899".parse::<SocketAddr>().unwrap());
-
-    let input = "asg";
-    assert!(parse_host_port(input.to_string(), 9123).is_err());
-    let input = "127.0.0:900";
-    assert!(parse_host_port(input.to_string(), 9123).is_err());
-    let input = "127.0.0";
-    assert!(parse_host_port(input.to_string(), 9123).is_err());
-    let input = "127.";
-    assert!(parse_host_port(input.to_string(), 9123).is_err());
-    let input = "127.9.0.1:asb";
-    assert!(parse_host_port(input.to_string(), 9123).is_err());
 }
 
 #[sim_test]
